+++ src/i18n/messages/ar.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "أضف محفظتي الخاصة",
   "WMe3nE": "استخدم EOA المنشئ المحدد أو قم بتكوين مفتاحه الخاص في متغيرات البيئة لتحديث القائمة البيضاء.",
   "4Ss2te": "دقة UMA غير متاحة مؤقتًا. يتم استخدام الحل المباشر حاليًا للأسواق الجديدة.",
+  "WFZl2u": "تعذر تحميل الفئات.",
+  "DLFcEW": "تعذر تحميل محافظ الخادم.",
+  "CBEgWt": "تعذر تحليل المسودة المحفوظة.",
+  "boL5xp": "فشل تحميل المسودة المحفوظة.",
+  "Jmi5-f": "تعذر حفظ شعار الفريق.",
+  "mB0ULq": "تعذر حفظ صورة الحدث.",
+  "KYMsGA": "تعذر حفظ صورة الخيار.",
+  "cpLRn-": "اجتاز فحص المحتوى بالذكاء الاصطناعي.",
+  "fe9NRj": "وجد فحص المحتوى بالذكاء الاصطناعي مشكلات.",
   "1C_yAq": "فشل التحقق من القائمة البيضاء للمقترح ({status})",
   "fLuAyG": "تعذر التحقق من صحة القائمة البيضاء لمقترحي الحلول.",
+  "UcK2Tu": "اكتملت جميع التوقيعات. سيكون الحدث الذي أنشأته متاحا على موقعك قريبا.",
+  "nWJM-J": "تمت استعادة تقدم التوقيع المعلق من الخادم.",
+  "5l0mkT": "تعذرت استعادة تقدم التوقيع المعلق.",
+  "Y5OzMI": "اكتملت المصادقة. لم يتم إرجاع أي معاملات للمنشئ.",
+  "iocMSm": "اكتملت المصادقة. تم تجهيز {txCount} طلبات توقيع.",
+  "YiKSfT": "اكتملت المصادقة. تم تجهيز طلب توقيع واحد ({txCount}).",
   "gVVoj8": "الموافقة على إنفاق USDC",
   "aljHil": "افتح محفظتك للسماح برسوم إنشاء السوق.",
   "A1MV7n": "ادفع رسوم الحل المباشر",
@@ -1115,6 +1130,11 @@
   "DKJULh": "افتح محفظتك لتفعيل التداول في هذا السوق.",
   "UmnM47": "تأكيد المعاملة",
   "-Mt_dr": "افتح محفظتك ووافق على استمرار المعاملة.",
+  "Pv3Fyn": "تم مسح النموذج.",
+  "88suad": "تعذر إكمال تدفق التوقيع.",
+  "vhUZ5B": "فشل إنشاء القواعد ({status})",
+  "UB_TJP": "تم إنشاء القواعد من {samplesUsed} عينات.",
+  "gEQgIO": "تعذر إنشاء القواعد بالذكاء الاصطناعي الآن.",
   "jr85XU": "فشل التحقق من المنشئين المسموح لهم ({status})",
   "d6Fypv": "تعذر التحقق من محافظ منشئي الأسواق المسموح لهم.",
   "w2hgle": "حدد محفظة EOA أولاً.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "فشل إضافة المنشئ المسموح له ({status})",
   "hXKzE3": "تمت إضافة المحفظة إلى منشئي الأسواق المسموح لهم.",
   "Bqnt0q": "تعذر إضافة المحفظة إلى منشئي الأسواق المسموح لهم.",
+  "WeVrut": "يلزم وجود خيارين على الأقل.",
+  "G4O7Qi": "يلزم وجود prop واحد على الأقل.",
+  "HAHS1f": "يلزم وجود صف سوق رياضي مخصص واحد على الأقل.",
   "-qMCrW": "إدارة رؤية الحدث وفحص الحجم والتحكم في كيفية نشر الأحداث المتزامنة الجديدة.",
   "vGhAxH": "غير مصادق عليه.",
   "TgfBJ4": "إدخال غير صالح.",
