+++ src/i18n/messages/de.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "meine eigene Wallet hinzufügen",
   "WMe3nE": "Verwenden Sie die EOA des ausgewählten Creators oder konfigurieren Sie dessen private key in den environment variables, um die Whitelist zu aktualisieren.",
   "4Ss2te": "UMA-Auflösung ist vorübergehend nicht verfügbar. Für neue Märkte wird derzeit direkte Auflösung verwendet.",
+  "WFZl2u": "Kategorien konnten nicht geladen werden.",
+  "DLFcEW": "Server-Wallets konnten nicht geladen werden.",
+  "CBEgWt": "Der gespeicherte Entwurf konnte nicht gelesen werden.",
+  "boL5xp": "Gespeicherter Entwurf konnte nicht geladen werden.",
+  "Jmi5-f": "Teamlogo konnte nicht gespeichert werden.",
+  "mB0ULq": "Eventbild konnte nicht gespeichert werden.",
+  "KYMsGA": "Optionsbild konnte nicht gespeichert werden.",
+  "cpLRn-": "KI-Inhaltsprüfung bestanden.",
+  "fe9NRj": "KI-Inhaltsprüfung hat Probleme gefunden.",
   "1C_yAq": "Prüfung der Proposer-Whitelist fehlgeschlagen ({status})",
   "fLuAyG": "Die Resolution-Proposer-Whitelist konnte nicht validiert werden.",
+  "UcK2Tu": "Alle Signaturen wurden abgeschlossen. Dein erstelltes Event wird in Kürze auf deiner Website verfügbar sein.",
+  "nWJM-J": "Ausstehender Signaturfortschritt vom Server wiederhergestellt.",
+  "5l0mkT": "Ausstehender Signaturfortschritt konnte nicht wiederhergestellt werden.",
+  "Y5OzMI": "Authentifizierung abgeschlossen. Es wurden keine Creator-Transaktionen zurückgegeben.",
+  "iocMSm": "Authentifizierung abgeschlossen. {txCount} Signaturanfragen vorbereitet.",
+  "YiKSfT": "Authentifizierung abgeschlossen. {txCount} Signaturanfrage vorbereitet.",
   "gVVoj8": "USDC-Ausgaben genehmigen",
   "aljHil": "Öffne deine Wallet, um die Gebühren für die Markterstellung zu erlauben.",
   "A1MV7n": "Direkte Auflösungsgebühr zahlen",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Öffne deine Wallet, um den Handel für diesen Markt zu aktivieren.",
   "UmnM47": "Transaktion bestätigen",
   "-Mt_dr": "Öffne deine Wallet und bestätige die Transaktion, um fortzufahren.",
+  "Pv3Fyn": "Formular geleert.",
+  "88suad": "Signaturablauf konnte nicht abgeschlossen werden.",
+  "vhUZ5B": "Regelgenerierung fehlgeschlagen ({status})",
+  "UB_TJP": "Regeln aus {samplesUsed} Beispielen generiert.",
+  "gEQgIO": "Regeln können gerade nicht mit KI generiert werden.",
   "jr85XU": "Prüfung der zugelassenen Ersteller fehlgeschlagen ({status})",
   "d6Fypv": "Wallets der zugelassenen Marktersteller konnten nicht validiert werden.",
   "w2hgle": "Wählen Sie zuerst eine EOA-Wallet aus.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Zugelassener Ersteller konnte nicht hinzugefügt werden ({status})",
   "hXKzE3": "Wallet wurde zu den zugelassenen Markterstellern hinzugefügt.",
   "Bqnt0q": "Wallet konnte nicht zu den zugelassenen Markterstellern hinzugefügt werden.",
+  "WeVrut": "Mindestens 2 Optionen sind erforderlich.",
+  "G4O7Qi": "Mindestens 1 Prop ist erforderlich.",
+  "HAHS1f": "Mindestens 1 benutzerdefinierte Sportmarktzeile ist erforderlich.",
   "-qMCrW": "Verwalte die Event-Sichtbarkeit, prüfe Volumenwerte und steuere, wie neue synchronisierte Events veröffentlicht werden.",
   "vGhAxH": "Nicht authentifiziert.",
   "TgfBJ4": "Ungültige Eingabe.",
