+++ src/i18n/messages/en.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "add my own wallet",
   "WMe3nE": "Use the selected creator EOA or configure its private key in environment variables to update the whitelist.",
   "4Ss2te": "UMA resolution is temporarily unavailable. Direct resolution is currently used for new markets.",
+  "WFZl2u": "Could not load categories.",
+  "DLFcEW": "Could not load server wallets.",
+  "CBEgWt": "The saved draft could not be parsed.",
+  "boL5xp": "Failed to load saved draft.",
+  "Jmi5-f": "Could not save team logo.",
+  "mB0ULq": "Could not save event image.",
+  "KYMsGA": "Could not save option image.",
+  "cpLRn-": "Content AI checker passed.",
+  "fe9NRj": "Content AI checker found issues.",
   "1C_yAq": "Proposer whitelist check failed ({status})",
   "fLuAyG": "Could not validate resolution proposers whitelist.",
+  "UcK2Tu": "All signatures completed. Your created event will be available on your site shortly.",
+  "nWJM-J": "Recovered pending signature progress from server.",
+  "5l0mkT": "Could not recover pending signature progress.",
+  "Y5OzMI": "Auth completed. No creator transactions were returned.",
+  "iocMSm": "Auth completed. Prepared {txCount} signature requests.",
+  "YiKSfT": "Auth completed. Prepared {txCount} signature request.",
   "gVVoj8": "Approve USDC spending",
   "aljHil": "Open your wallet to allow the market creation fees.",
   "A1MV7n": "Pay direct resolution fee",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Open your wallet to activate trading for this market.",
   "UmnM47": "Confirm transaction",
   "-Mt_dr": "Open your wallet and approve the transaction to continue.",
+  "Pv3Fyn": "Form cleared.",
+  "88suad": "Could not complete signature flow.",
+  "vhUZ5B": "Rules generation failed ({status})",
+  "UB_TJP": "Rules generated from {samplesUsed} samples.",
+  "gEQgIO": "Could not generate rules with AI right now.",
   "jr85XU": "Allowed creators check failed ({status})",
   "d6Fypv": "Could not validate allowed market creator wallets.",
   "w2hgle": "Select an EOA wallet first.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Failed to add allowed creator ({status})",
   "hXKzE3": "Wallet added to allowed market creator wallets.",
   "Bqnt0q": "Could not add wallet to allowed market creator wallets.",
+  "WeVrut": "At least 2 options are required.",
+  "G4O7Qi": "At least 1 prop is required.",
+  "HAHS1f": "At least 1 custom sports market row is required.",
   "-qMCrW": "Manage event visibility, inspect volume, and control how new synced events are deployed.",
   "vGhAxH": "Unauthenticated.",
   "TgfBJ4": "Invalid input.",
