+++ src/i18n/messages/es.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "añadir mi propia wallet",
   "WMe3nE": "Usa la EOA del creator seleccionado o configura su private key en las environment variables para actualizar la whitelist.",
   "4Ss2te": "La resolución por UMA no está disponible temporalmente. Actualmente se usa resolución directa para nuevos mercados.",
+  "WFZl2u": "No se pudieron cargar las categorías.",
+  "DLFcEW": "No se pudieron cargar las wallets del servidor.",
+  "CBEgWt": "No se pudo interpretar el borrador guardado.",
+  "boL5xp": "No se pudo cargar el borrador guardado.",
+  "Jmi5-f": "No se pudo guardar el logo del equipo.",
+  "mB0ULq": "No se pudo guardar la imagen del evento.",
+  "KYMsGA": "No se pudo guardar la imagen de la opción.",
+  "cpLRn-": "La comprobación de contenido con IA se aprobó.",
+  "fe9NRj": "La comprobación de contenido con IA encontró problemas.",
   "1C_yAq": "Error al comprobar la whitelist de proposers ({status})",
   "fLuAyG": "No se pudo validar la whitelist de proposers de resolución.",
+  "UcK2Tu": "Todas las firmas se completaron. El evento creado estará disponible en tu sitio pronto.",
+  "nWJM-J": "Progreso de firma pendiente recuperado del servidor.",
+  "5l0mkT": "No se pudo recuperar el progreso de firma pendiente.",
+  "Y5OzMI": "Autenticación completada. No se devolvieron transacciones del creador.",
+  "iocMSm": "Autenticación completada. Se prepararon {txCount} solicitudes de firma.",
+  "YiKSfT": "Autenticación completada. Se preparó {txCount} solicitud de firma.",
   "gVVoj8": "Aprobar gasto de USDC",
   "aljHil": "Abre tu wallet para permitir las fees de creación del mercado.",
   "A1MV7n": "Pagar fee de resolución directa",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Abre tu wallet para activar el trading de este mercado.",
   "UmnM47": "Confirmar transacción",
   "-Mt_dr": "Abre tu wallet y aprueba la transacción para continuar.",
+  "Pv3Fyn": "Formulario borrado.",
+  "88suad": "No se pudo completar el flujo de firma.",
+  "vhUZ5B": "La generación de reglas falló ({status})",
+  "UB_TJP": "Reglas generadas a partir de {samplesUsed} muestras.",
+  "gEQgIO": "No se pueden generar reglas con IA ahora.",
   "jr85XU": "La comprobación de creadores permitidos falló ({status})",
   "d6Fypv": "No se pudieron validar las wallets de creadores de mercado permitidos.",
   "w2hgle": "Selecciona primero una wallet EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "No se pudo añadir el creador permitido ({status})",
   "hXKzE3": "Wallet añadida a los creadores de mercado permitidos.",
   "Bqnt0q": "No se pudo añadir la wallet a los creadores de mercado permitidos.",
+  "WeVrut": "Se requieren al menos 2 opciones.",
+  "G4O7Qi": "Se requiere al menos 1 prop.",
+  "HAHS1f": "Se requiere al menos 1 fila personalizada de mercado deportivo.",
   "-qMCrW": "Administra la visibilidad de los eventos, revisa el volumen y controla cómo se publican los nuevos eventos sincronizados.",
   "vGhAxH": "No autenticado.",
   "TgfBJ4": "Entrada no válida.",
