+++ src/i18n/messages/fr.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "ajouter ma propre wallet",
   "WMe3nE": "Utilisez l'EOA du creator sélectionné ou configurez sa private key dans les environment variables pour mettre à jour la whitelist.",
   "4Ss2te": "La résolution UMA est temporairement indisponible. La résolution directe est actuellement utilisée pour les nouveaux marchés.",
+  "WFZl2u": "Impossible de charger les catégories.",
+  "DLFcEW": "Impossible de charger les wallets serveur.",
+  "CBEgWt": "Le brouillon enregistré n'a pas pu être analysé.",
+  "boL5xp": "Échec du chargement du brouillon enregistré.",
+  "Jmi5-f": "Impossible d'enregistrer le logo de l'équipe.",
+  "mB0ULq": "Impossible d'enregistrer l'image de l'événement.",
+  "KYMsGA": "Impossible d'enregistrer l'image de l'option.",
+  "cpLRn-": "La vérification du contenu par IA a réussi.",
+  "fe9NRj": "La vérification du contenu par IA a trouvé des problèmes.",
   "1C_yAq": "Échec de la vérification de la whitelist des proposers ({status})",
   "fLuAyG": "Impossible de valider la whitelist des proposers de résolution.",
+  "UcK2Tu": "Toutes les signatures sont terminées. L'événement créé sera bientôt disponible sur votre site.",
+  "nWJM-J": "Progression de signature en attente récupérée depuis le serveur.",
+  "5l0mkT": "Impossible de récupérer la progression de signature en attente.",
+  "Y5OzMI": "Authentification terminée. Aucune transaction du créateur n'a été renvoyée.",
+  "iocMSm": "Authentification terminée. {txCount} demandes de signature préparées.",
+  "YiKSfT": "Authentification terminée. {txCount} demande de signature préparée.",
   "gVVoj8": "Approuver la dépense USDC",
   "aljHil": "Ouvrez votre wallet pour autoriser les frais de création du marché.",
   "A1MV7n": "Payer les frais de résolution directe",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Ouvrez votre wallet pour activer le trading sur ce marché.",
   "UmnM47": "Confirmer la transaction",
   "-Mt_dr": "Ouvrez votre wallet et approuvez la transaction pour continuer.",
+  "Pv3Fyn": "Formulaire effacé.",
+  "88suad": "Impossible de terminer le flux de signature.",
+  "vhUZ5B": "La génération des règles a échoué ({status})",
+  "UB_TJP": "Règles générées à partir de {samplesUsed} échantillons.",
+  "gEQgIO": "Impossible de générer des règles avec l'IA pour le moment.",
   "jr85XU": "La vérification des créateurs autorisés a échoué ({status})",
   "d6Fypv": "Impossible de valider les wallets des créateurs de marché autorisés.",
   "w2hgle": "Sélectionnez d'abord une wallet EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Échec de l’ajout du créateur autorisé ({status})",
   "hXKzE3": "Wallet ajouté aux créateurs de marché autorisés.",
   "Bqnt0q": "Impossible d’ajouter le wallet aux créateurs de marché autorisés.",
+  "WeVrut": "Au moins 2 options sont requises.",
+  "G4O7Qi": "Au moins 1 prop est requis.",
+  "HAHS1f": "Au moins 1 ligne personnalisée de marché sportif est requise.",
   "-qMCrW": "Gérez la visibilité des événements, consultez le volume et contrôlez la publication des nouveaux événements synchronisés.",
   "vGhAxH": "Non authentifié.",
   "TgfBJ4": "Entrée invalide.",
