+++ src/i18n/messages/it.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "aggiungi il mio portafoglio",
   "WMe3nE": "Utilizza l'EOA del creatore selezionato o configura la sua chiave privata nelle variabili di ambiente per aggiornare la whitelist.",
   "4Ss2te": "La risoluzione UMA è temporaneamente non disponibile. La risoluzione diretta è attualmente utilizzata per i nuovi mercati.",
+  "WFZl2u": "Impossibile caricare le categorie.",
+  "DLFcEW": "Impossibile caricare i wallet del server.",
+  "CBEgWt": "La bozza salvata non può essere interpretata.",
+  "boL5xp": "Impossibile caricare la bozza salvata.",
+  "Jmi5-f": "Impossibile salvare il logo della squadra.",
+  "mB0ULq": "Impossibile salvare l'immagine dell'evento.",
+  "KYMsGA": "Impossibile salvare l'immagine dell'opzione.",
+  "cpLRn-": "Controllo dei contenuti IA superato.",
+  "fe9NRj": "Il controllo dei contenuti IA ha trovato problemi.",
   "1C_yAq": "Controllo della lista bianca del proponente non riuscito ({status})",
   "fLuAyG": "Impossibile convalidare la lista bianca dei proponenti della risoluzione.",
+  "UcK2Tu": "Tutte le firme sono state completate. L'evento creato sarà presto disponibile sul tuo sito.",
+  "nWJM-J": "Avanzamento della firma in sospeso recuperato dal server.",
+  "5l0mkT": "Impossibile recuperare l'avanzamento della firma in sospeso.",
+  "Y5OzMI": "Autenticazione completata. Non sono state restituite transazioni del creator.",
+  "iocMSm": "Autenticazione completata. Preparata/e {txCount} richieste di firma.",
+  "YiKSfT": "Autenticazione completata. Preparata {txCount} richiesta di firma.",
   "gVVoj8": "Approvare la spesa dell'USDC",
   "aljHil": "Apri il tuo portafoglio per consentire le commissioni di creazione del mercato.",
   "A1MV7n": "Pagare una commissione di risoluzione diretta",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Apri il tuo portafoglio per attivare il trading per questo mercato.",
   "UmnM47": "Conferma la transazione",
   "-Mt_dr": "Apri il tuo portafoglio e approva la transazione per continuare.",
+  "Pv3Fyn": "Modulo cancellato.",
+  "88suad": "Impossibile completare il flusso di firma.",
+  "vhUZ5B": "Generazione delle regole non riuscita ({status})",
+  "UB_TJP": "Regole generate da {samplesUsed} campioni.",
+  "gEQgIO": "Impossibile generare regole con l'IA in questo momento.",
   "jr85XU": "Controllo dei creatori consentiti non riuscito ({status})",
   "d6Fypv": "Impossibile convalidare i wallet dei creatori di mercato consentiti.",
   "w2hgle": "Seleziona prima un portafoglio EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Impossibile aggiungere il creatore consentito ({status})",
   "hXKzE3": "Wallet aggiunto ai creatori di mercato consentiti.",
   "Bqnt0q": "Impossibile aggiungere il wallet ai creatori di mercato consentiti.",
+  "WeVrut": "Sono richieste almeno 2 opzioni.",
+  "G4O7Qi": "È richiesto almeno 1 prop.",
+  "HAHS1f": "È richiesta almeno 1 riga personalizzata del mercato sportivo.",
   "-qMCrW": "Gestisci la visibilità degli eventi, controlla il volume e controlla il modo in cui vengono distribuiti i nuovi eventi sincronizzati.",
   "vGhAxH": "Non autenticato.",
   "TgfBJ4": "Immissione non valida.",
