+++ src/i18n/messages/ja.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "自分のウォレットを追加する",
   "WMe3nE": "選択した作成者 EOA を使用するか、環境変数でその秘密キーを構成してホワイトリストを更新します。",
   "4Ss2te": "UMA 解決は一時的に利用できません。直接解決は現在、新しい市場に使用されています。",
+  "WFZl2u": "カテゴリを読み込めませんでした。",
+  "DLFcEW": "サーバーウォレットを読み込めませんでした。",
+  "CBEgWt": "保存済みの下書きを解析できませんでした。",
+  "boL5xp": "保存済みの下書きを読み込めませんでした。",
+  "Jmi5-f": "チームロゴを保存できませんでした。",
+  "mB0ULq": "イベント画像を保存できませんでした。",
+  "KYMsGA": "オプション画像を保存できませんでした。",
+  "cpLRn-": "AIコンテンツチェックに合格しました。",
+  "fe9NRj": "AIコンテンツチェックで問題が見つかりました。",
   "1C_yAq": "プロポーザーのホワイトリストのチェックに失敗しました ({status})",
   "fLuAyG": "決議提案者のホワイトリストを検証できませんでした。",
+  "UcK2Tu": "すべての署名が完了しました。作成したイベントはまもなくサイトで利用可能になります。",
+  "nWJM-J": "保留中の署名進行状況をサーバーから復元しました。",
+  "5l0mkT": "保留中の署名進行状況を復元できませんでした。",
+  "Y5OzMI": "認証が完了しました。作成者のトランザクションは返されませんでした。",
+  "iocMSm": "認証が完了しました。{txCount}件の署名リクエストを準備しました。",
+  "YiKSfT": "認証が完了しました。{txCount}件の署名リクエストを準備しました。",
   "gVVoj8": "USDC の支出を承認する",
   "aljHil": "ウォレットを開いてマーケット作成手数料を許可します。",
   "A1MV7n": "解決手数料を直接支払う",
@@ -1115,6 +1130,11 @@
   "DKJULh": "ウォレットを開いて、この市場の取引をアクティブにします。",
   "UmnM47": "取引の確認",
   "-Mt_dr": "ウォレットを開いてトランザクションを承認して続行します。",
+  "Pv3Fyn": "フォームをクリアしました。",
+  "88suad": "署名フローを完了できませんでした。",
+  "vhUZ5B": "ルール生成に失敗しました（{status}）",
+  "UB_TJP": "{samplesUsed}件のサンプルからルールを生成しました。",
+  "gEQgIO": "現在、AIでルールを生成できません。",
   "jr85XU": "許可された作成者の確認に失敗しました（{status}）",
   "d6Fypv": "許可されたマーケット作成者のウォレットを検証できませんでした。",
   "w2hgle": "最初に EOA ウォレットを選択します。",
@@ -1122,6 +1142,9 @@
   "d-msWg": "許可された作成者を追加できませんでした（{status}）",
   "hXKzE3": "ウォレットを許可されたマーケット作成者に追加しました。",
   "Bqnt0q": "ウォレットを許可されたマーケット作成者に追加できませんでした。",
+  "WeVrut": "少なくとも2つのオプションが必要です。",
+  "G4O7Qi": "少なくとも1つのpropが必要です。",
+  "HAHS1f": "少なくとも1つのカスタムスポーツ市場行が必要です。",
   "-qMCrW": "イベントの可視性を管理し、ボリュームを検査し、新しい同期イベントの展開方法を制御します。",
   "vGhAxH": "未認証。",
   "TgfBJ4": "入力が無効です。",
