+++ src/i18n/messages/pl.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "dodaj mój własny portfel",
   "WMe3nE": "Użyj wybranego EOA twórcy lub skonfiguruj jego klucz prywatny w zmiennych środowiskowych, aby zaktualizować białą listę.",
   "4Ss2te": "Rozdzielczość UMA jest tymczasowo niedostępna. Rozdzielczość bezpośrednia jest obecnie stosowana w przypadku nowych rynków.",
+  "WFZl2u": "Nie udało się wczytać kategorii.",
+  "DLFcEW": "Nie udało się wczytać portfeli serwera.",
+  "CBEgWt": "Nie można było przetworzyć zapisanego szkicu.",
+  "boL5xp": "Nie udało się wczytać zapisanego szkicu.",
+  "Jmi5-f": "Nie udało się zapisać logo drużyny.",
+  "mB0ULq": "Nie udało się zapisać obrazu wydarzenia.",
+  "KYMsGA": "Nie udało się zapisać obrazu opcji.",
+  "cpLRn-": "Kontrola treści przez AI zakończona pomyślnie.",
+  "fe9NRj": "Kontrola treści przez AI znalazła problemy.",
   "1C_yAq": "Sprawdzenie białej listy wnioskodawców nie powiodło się ({status})",
   "fLuAyG": "Nie można sprawdzić białej listy propozycji rozwiązań.",
+  "UcK2Tu": "Wszystkie podpisy zostały ukończone. Utworzone wydarzenie będzie wkrótce dostępne na Twojej stronie.",
+  "nWJM-J": "Odzyskano z serwera oczekujący postęp podpisu.",
+  "5l0mkT": "Nie udało się odzyskać oczekującego postępu podpisu.",
+  "Y5OzMI": "Uwierzytelnianie zakończone. Nie zwrócono żadnych transakcji twórcy.",
+  "iocMSm": "Uwierzytelnianie zakończone. Przygotowano {txCount} żądań podpisu.",
+  "YiKSfT": "Uwierzytelnianie zakończone. Przygotowano {txCount} żądanie podpisu.",
   "gVVoj8": "Zatwierdź wydatki USDC",
   "aljHil": "Otwórz portfel, aby zezwolić na opłaty za tworzenie rynku.",
   "A1MV7n": "Zapłać bezpośrednią opłatę za rozstrzygnięcie",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Otwórz portfel, aby aktywować handel na tym rynku.",
   "UmnM47": "Potwierdź transakcję",
   "-Mt_dr": "Otwórz portfel i zatwierdź transakcję, aby kontynuować.",
+  "Pv3Fyn": "Formularz wyczyszczony.",
+  "88suad": "Nie udało się ukończyć przepływu podpisu.",
+  "vhUZ5B": "Generowanie reguł nie powiodło się ({status})",
+  "UB_TJP": "Reguły wygenerowano na podstawie {samplesUsed} próbek.",
+  "gEQgIO": "Nie można teraz wygenerować reguł za pomocą AI.",
   "jr85XU": "Sprawdzanie dozwolonych twórców nie powiodło się ({status})",
   "d6Fypv": "Nie udało się zweryfikować portfeli dozwolonych twórców rynków.",
   "w2hgle": "Najpierw wybierz portfel EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Nie udało się dodać dozwolonego twórcy ({status})",
   "hXKzE3": "Portfel dodano do dozwolonych twórców rynków.",
   "Bqnt0q": "Nie udało się dodać portfela do dozwolonych twórców rynków.",
+  "WeVrut": "Wymagane są co najmniej 2 opcje.",
+  "G4O7Qi": "Wymagany jest co najmniej 1 prop.",
+  "HAHS1f": "Wymagany jest co najmniej 1 niestandardowy wiersz rynku sportowego.",
   "-qMCrW": "Zarządzaj widocznością zdarzeń, sprawdzaj ich liczbę i kontroluj sposób wdrażania nowych zsynchronizowanych zdarzeń.",
   "vGhAxH": "Nieuwierzytelnione.",
   "TgfBJ4": "Nieprawidłowe dane wejściowe.",
