+++ src/i18n/messages/pt.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "adicionar minha própria wallet",
   "WMe3nE": "Use a EOA do creator selecionado ou configure sua private key nas environment variables para atualizar a whitelist.",
   "4Ss2te": "A resolução via UMA está temporariamente indisponível. A resolução direta está sendo usada para novos mercados.",
+  "WFZl2u": "Não foi possível carregar as categorias.",
+  "DLFcEW": "Não foi possível carregar as wallets do servidor.",
+  "CBEgWt": "Não foi possível interpretar o rascunho salvo.",
+  "boL5xp": "Falha ao carregar o rascunho salvo.",
+  "Jmi5-f": "Não foi possível salvar o logo do time.",
+  "mB0ULq": "Não foi possível salvar a imagem do evento.",
+  "KYMsGA": "Não foi possível salvar a imagem da opção.",
+  "cpLRn-": "Verificação de conteúdo por IA aprovada.",
+  "fe9NRj": "A verificação de conteúdo por IA encontrou problemas.",
   "1C_yAq": "Falha ao verificar whitelist de proposers ({status})",
   "fLuAyG": "Não foi possível validar a whitelist de proposers de resolução.",
+  "UcK2Tu": "Todas as assinaturas foram concluídas. Seu evento criado estará disponível no site em breve.",
+  "nWJM-J": "Progresso de assinatura pendente recuperado do servidor.",
+  "5l0mkT": "Não foi possível recuperar o progresso de assinatura pendente.",
+  "Y5OzMI": "Autenticação concluída. Nenhuma transação do creator foi retornada.",
+  "iocMSm": "Autenticação concluída. Preparadas {txCount} solicitações de assinatura.",
+  "YiKSfT": "Autenticação concluída. Preparada {txCount} solicitação de assinatura.",
   "gVVoj8": "Aprovar gasto de USDC",
   "aljHil": "Abra sua wallet para permitir as fees de criação do mercado.",
   "A1MV7n": "Pagar fee de resolução direta",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Abra sua wallet para ativar o trading deste mercado.",
   "UmnM47": "Confirmar transação",
   "-Mt_dr": "Abra sua wallet e aprove a transação para continuar.",
+  "Pv3Fyn": "Formulário limpo.",
+  "88suad": "Não foi possível concluir o fluxo de assinatura.",
+  "vhUZ5B": "Falha ao gerar regras ({status})",
+  "UB_TJP": "Regras geradas a partir de {samplesUsed} amostras.",
+  "gEQgIO": "Não foi possível gerar regras com IA agora.",
   "jr85XU": "A verificação de criadores permitidos falhou ({status})",
   "d6Fypv": "Não foi possível validar as wallets de criadores de mercado permitidos.",
   "w2hgle": "Selecione primeiro uma wallet EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Falha ao adicionar criador permitido ({status})",
   "hXKzE3": "Wallet adicionada aos criadores de mercado permitidos.",
   "Bqnt0q": "Não foi possível adicionar a wallet aos criadores de mercado permitidos.",
+  "WeVrut": "Pelo menos 2 opções são obrigatórias.",
+  "G4O7Qi": "Pelo menos 1 prop é obrigatório.",
+  "HAHS1f": "Pelo menos 1 linha personalizada de mercado esportivo é obrigatória.",
   "-qMCrW": "Gerencie a visibilidade dos eventos, veja o volume e controle como novos eventos sincronizados são publicados.",
   "vGhAxH": "Não autenticado.",
   "TgfBJ4": "Entrada inválida.",
