+++ src/i18n/messages/ru.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "добавить свой кошелек",
   "WMe3nE": "Используйте EOA выбранного создателя или настройте его закрытый ключ в переменных среды, чтобы обновить белый список.",
   "4Ss2te": "Разрешение UMA временно недоступно. Прямое разрешение в настоящее время используется для новых рынков.",
+  "WFZl2u": "Не удалось загрузить категории.",
+  "DLFcEW": "Не удалось загрузить серверные кошельки.",
+  "CBEgWt": "Не удалось разобрать сохраненный черновик.",
+  "boL5xp": "Не удалось загрузить сохраненный черновик.",
+  "Jmi5-f": "Не удалось сохранить логотип команды.",
+  "mB0ULq": "Не удалось сохранить изображение события.",
+  "KYMsGA": "Не удалось сохранить изображение опции.",
+  "cpLRn-": "Проверка контента ИИ пройдена.",
+  "fe9NRj": "Проверка контента ИИ обнаружила проблемы.",
   "1C_yAq": "Не удалось проверить белый список предлагающих ({status})",
   "fLuAyG": "Не удалось проверить белый список авторов решений.",
+  "UcK2Tu": "Все подписи завершены. Созданное событие скоро будет доступно на вашем сайте.",
+  "nWJM-J": "Ожидающий прогресс подписи восстановлен с сервера.",
+  "5l0mkT": "Не удалось восстановить ожидающий прогресс подписи.",
+  "Y5OzMI": "Аутентификация завершена. Транзакции создателя не были возвращены.",
+  "iocMSm": "Аутентификация завершена. Подготовлено запросов подписи: {txCount}.",
+  "YiKSfT": "Аутентификация завершена. Подготовлен запрос подписи: {txCount}.",
   "gVVoj8": "Утвердить расходы USDC",
   "aljHil": "Откройте свой кошелек, чтобы разрешить комиссию за создание рынка.",
   "A1MV7n": "Оплатить прямую комиссию за разрешение",
@@ -1115,6 +1130,11 @@
   "DKJULh": "Откройте свой кошелек, чтобы активировать торговлю на этом рынке.",
   "UmnM47": "Подтвердить транзакцию",
   "-Mt_dr": "Откройте свой кошелек и подтвердите транзакцию, чтобы продолжить.",
+  "Pv3Fyn": "Форма очищена.",
+  "88suad": "Не удалось завершить процесс подписи.",
+  "vhUZ5B": "Не удалось сгенерировать правила ({status})",
+  "UB_TJP": "Правила сгенерированы на основе {samplesUsed} примеров.",
+  "gEQgIO": "Сейчас не удалось сгенерировать правила с помощью ИИ.",
   "jr85XU": "Не удалось проверить разрешенных создателей ({status})",
   "d6Fypv": "Не удалось проверить кошельки разрешенных создателей рынков.",
   "w2hgle": "Сначала выберите кошелек EOA.",
@@ -1122,6 +1142,9 @@
   "d-msWg": "Не удалось добавить разрешенного создателя ({status})",
   "hXKzE3": "Кошелек добавлен в список разрешенных создателей рынков.",
   "Bqnt0q": "Не удалось добавить кошелек в список разрешенных создателей рынков.",
+  "WeVrut": "Требуется как минимум 2 опции.",
+  "G4O7Qi": "Требуется как минимум 1 prop.",
+  "HAHS1f": "Требуется как минимум 1 пользовательская строка спортивного рынка.",
   "-qMCrW": "Управляйте видимостью событий, проверяйте объем и контролируйте развертывание новых синхронизированных событий.",
   "vGhAxH": "Неаутентифицированный.",
   "TgfBJ4": "Неверный ввод.",
