+++ src/i18n/messages/zh.json
@@ -1103,8 +1103,23 @@
   "Uei2Lk": "添加我自己的钱包",
   "WMe3nE": "请使用所选 creator 的 EOA，或在 environment variables 中配置其 private key，以更新 whitelist。",
   "4Ss2te": "UMA 结算暂时不可用。新市场目前使用直接结算。",
+  "WFZl2u": "无法加载分类。",
+  "DLFcEW": "无法加载服务器钱包。",
+  "CBEgWt": "无法解析已保存的草稿。",
+  "boL5xp": "无法加载已保存的草稿。",
+  "Jmi5-f": "无法保存球队徽标。",
+  "mB0ULq": "无法保存事件图片。",
+  "KYMsGA": "无法保存选项图片。",
+  "cpLRn-": "AI 内容检查已通过。",
+  "fe9NRj": "AI 内容检查发现问题。",
   "1C_yAq": "Proposer whitelist 检查失败（{status}）",
   "fLuAyG": "无法验证 resolution proposers whitelist。",
+  "UcK2Tu": "所有签名已完成。你创建的事件很快将在你的网站上可用。",
+  "nWJM-J": "已从服务器恢复待处理的签名进度。",
+  "5l0mkT": "无法恢复待处理的签名进度。",
+  "Y5OzMI": "认证已完成。未返回创建者交易。",
+  "iocMSm": "认证已完成。已准备 {txCount} 个签名请求。",
+  "YiKSfT": "认证已完成。已准备 {txCount} 个签名请求。",
   "gVVoj8": "批准 USDC 支出",
   "aljHil": "打开你的钱包以授权市场创建费用。",
   "A1MV7n": "支付直接结算费用",
@@ -1115,6 +1130,11 @@
   "DKJULh": "打开你的钱包以启用该市场的交易。",
   "UmnM47": "确认交易",
   "-Mt_dr": "打开你的钱包并批准交易以继续。",
+  "Pv3Fyn": "表单已清空。",
+  "88suad": "无法完成签名流程。",
+  "vhUZ5B": "规则生成失败（{status}）",
+  "UB_TJP": "已根据 {samplesUsed} 个样本生成规则。",
+  "gEQgIO": "目前无法使用 AI 生成规则。",
   "jr85XU": "允许创建者检查失败（{status}）",
   "d6Fypv": "无法验证允许的市场创建者钱包。",
   "w2hgle": "请先选择一个 EOA 钱包。",
@@ -1122,6 +1142,9 @@
   "d-msWg": "无法添加允许的创建者（{status}）",
   "hXKzE3": "钱包已添加到允许的市场创建者。",
   "Bqnt0q": "无法将钱包添加到允许的市场创建者。",
+  "WeVrut": "至少需要 2 个选项。",
+  "G4O7Qi": "至少需要 1 个 prop。",
+  "HAHS1f": "至少需要 1 行自定义体育市场。",
   "-qMCrW": "管理事件可见性、查看成交量，并控制新同步事件的发布方式。",
   "vGhAxH": "未认证。",
   "TgfBJ4": "输入无效。",
